requirements.txt
numpy
pandas
fave
praat-parselmouth
pympi-ling
